requirements.txt
fire
torch
bitsandbytes
accelerate
transformers
scipy
fastapi
uvicorn
pyyaml
